# documents/edificad-swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Edificad API",
    "description": "Espeicificação da API Edificad, voltada para a gestão da ONG Edificando Vidas. ",
    "version": "0.0.1",
    "contact": {
      "name": "Repositório",
      "url": "https://github.com/jarnunes/edificad"
    },
    "license": {
      "name": "GPLv3",
      "url": "https://www.gnu.org/licenses/gpl-3.0.en.html"
    }
  },
  "servers": [
    {
      "url": "https://edificad-production.up.railway.app/api",
      "description": "Produção"
    },
    {
      "url": "http://edificad.com:5000/api",
      "description": "Desenvolvimento"
    }
  ],
  "paths": {
...
    "/beneficiario": {
      "get": {
        "tags": [
          "Beneficiário"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "nome",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cpf",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BeneficiarioArray"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Beneficiário"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BeneficiarioPost"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Beneficiario"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Beneficiário"
        ],
        "summary": "Atualizar entidade",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Beneficiario"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Beneficiario"
                }
              }
            }
          }
        }
      }
    },
    "/beneficiario/{id}": {
      "get": {
        "tags": [
          "Beneficiário"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID da entidade",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Beneficiario"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Beneficiário"
        ],
        "summary": "Remove um beneficiário referente ao ID do parâmetro",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do beneficiário a ser removido",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Beneficiario"
                }
              }
            }
          },
          "404": {
            "description": "Não encontrado",
            "content": {
              "application/json": {
                "example": "Objeto não encontrado"
              }
            }
          }
        }
      }
    },
    "/beneficiario/create-from-list": {
      "post": {
        "tags": [
          "Beneficiário"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BeneficiarioPostArray"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BeneficiarioArray"
                }
              }
            }
          }
        }
      }
    },
    "/cesta": {
      "get": {
        "tags": [
          "Cesta"
        ],
        "summary": "Retornar array com os itens cadastrados",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "nome",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CestaArray"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Cesta"
        ],
        "summary": "Cadastrar entidade",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CestaPost"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cesta"
                }
              }
            }
          },
          "422": {
            "description": "Erro de validação",
            "content": {
              "application/json": {
                "example": "Ocorreu um erro na validação do campo 'nome_campo'"
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Cesta"
        ],
        "summary": "Atualizar entidade",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Cesta"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cesta"
                }
              }
            }
          }
        }
      }
    },
    "/cesta/{id}": {
      "delete": {
        "tags": [
          "Cesta"
        ],
        "summary": "Remove a cesta referente ao ID do parâmetro",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID da cesta a ser removida",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "example": "Cesta {nome} removida com sucesso."
              }
            }
          },
          "401": {
            "description": "Não Autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenInvalido"
                }
              }
            }
          },
          "404": {
            "description": "Não encontrado",
            "content": {
              "application/json": {
                "example": "Objeto não encontrado"
              }
            }
          }
        }
      }
    },
    "/cesta/create-from-list": {
      "post": {
        "tags": [
          "Cesta"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CestaPostArray"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CestaArray"
                }
              }
            }
          }
        }
      }
    },
...
    "/distribuicao-cesta": {
      "get": {
        "tags": [
          "Distribuição Cesta"
        ],
        "summary": "Retorna um array com todas as distribuições registradas",
        "parameters": [
          {
            "name": "cesta",
            "in": "query",
            "description": "Nome da cesta",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cpfBeneficiario",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cpfVoluntario",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "data",
            "in": "query",
            "description": "Data de distribuição da cesta",
            "schema": {
              "type": "string",
              "example": "2023-10-28"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DistribuicaoCestaArray"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Distribuição Cesta"
        ],
        "summary": "Registar nova distribuição de cesta",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DistribuicaoCestaPost"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DistribuicaoCesta"
                }
              }
            }
          }
        }
      }
    },
...
    "/relatorios/resumo-beneficiarios-assistidos": {
      "get": {
        "tags": [
          "Relatórios"
        ],
        "parameters": [
          {
            "name": "anoReferencia",
            "in": "query",
            "description": "Ano de Referência da consulta",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 2023
            }
          },
          {
            "name": "mesReferencia",
            "in": "query",
            "description": "Mês de referencia da consulta",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 7
            }
          },
          {
            "name": "qtdMesesAnteriores",
            "in": "query",
            "description": "Quantidade de meses anteriores ao mes de referência para constituição de um período consulta",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 12
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumoQuantidadesPorMesAnoArray"
                }
              }
            }
          }
        }
      }
    },
    "/relatorios/resumo-cestas-distribuidas": {
      "get": {
        "tags": [
          "Relatórios"
        ],
        "parameters": [
          {
            "name": "anoReferencia",
            "in": "query",
            "description": "Ano de Referência da consulta",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 2023
            }
          },
          {
            "name": "mesReferencia",
            "in": "query",
            "description": "Mês de referencia da consulta",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 7
            }
          },
          {
            "name": "qtdMesesAnteriores",
            "in": "query",
            "description": "Quantidade de meses anteriores ao mes de referência para constituição de um período consulta",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 12
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumoQuantidadesPorMesAnoArray"
                }
              }
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "auth": []
    }
  ],
  "components": {
    "schemas": {
...
      "BeneficiarioPost": {
        "type": "object",
        "properties": {
          "nome": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "example": "beneficiario@gmail.com"
          },
          "cpf": {
            "type": "string",
            "example": "000.000.000-00"
          },
          "telefone": {
            "type": "string",
            "example": "(99) 9 9999-9999"
          },
          "data_nascimento": {
            "type": "string",
            "example": "yyyy-mm-dd"
          },
          "endereco": {
            "$ref": "#/components/schemas/EnderecoPost"
          }
        }
      },
      "BeneficiarioPostArray": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/BeneficiarioPost"
        }
      },
      "Beneficiario": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "nome": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "example": "beneficiario@gmail.com"
          },
          "cpf": {
            "type": "string",
            "example": "000.000.000-00"
          },
          "telefone": {
            "type": "string",
            "example": "(99) 9 9999-9999"
          },
          "data_nascimento": {
            "type": "string",
            "example": "yyyy-mm-dd"
          },
          "endereco": {
            "$ref": "#/components/schemas/Endereco"
          },
          "dependentes": {
            "$ref": "#/components/schemas/DependenteArray"
          }
        }
      },
      "BeneficiarioArray": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/Beneficiario"
        }
      },
      "BeneficiarioDistribuicaoCesta": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 999
          },
          "nome": {
            "type": "string",
            "example": "Beneficiário A"
          },
          "email": {
            "type": "string",
            "example": "beneficiario@email.com"
          },
          "cpf": {
            "type": "string",
            "example": "000.000.000-00"
          },
          "telefone": {
            "type": "string",
            "example": "(99) 99999-9999"
          },
          "data_nascimento": {
            "type": "string",
            "example": "yyyy-mm-dd"
          },
          "qtde_projetos_participado": {
            "type": "integer",
            "example": 5
          }
        }
      },
      "Cesta": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "nome": {
            "type": "string",
            "example": "Cesta AB"
          },
          "descricao": {
            "type": "string",
            "example": "Destinada a familias com 4 a 6 integrantes"
          },
          "quantidade_estoque": {
            "type": "integer",
            "example": 60
          }
        }
      },
      "CestaArray": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/Cesta"
        }
      },
      "CestaPost": {
        "type": "object",
        "properties": {
          "nome": {
            "type": "string"
          },
          "descricao": {
            "type": "string"
          },
          "quantidade_estoque": {
            "type": "integer"
          }
        }
      },
      "CestaPostArray": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/CestaPost"
        }
      },
      "CestaDistribuicao": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 999
          },
          "nome": {
            "type": "string",
            "example": "CESTA 01"
          }
        }
      },
...
      "VoluntarioDistribuicaoCesta": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 999
          },
          "nome": {
            "type": "string",
            "example": "Voluntario A"
          },
          "email": {
            "type": "string",
            "example": "voluntario@email.com"
          },
          "cpf": {
            "type": "string",
            "example": "000.000.000-00"
          },
          "telefone": {
            "type": "string",
            "example": "(99) 99999-9999"
          },
          "data_nascimento": {
            "type": "string",
            "example": "yyyy-mm-dd"
          },
          "qtde_projetos_participado": {
            "type": "integer",
            "example": 5
          }
        }
      },
      "Dependente": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "nome": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "cpf": {
            "type": "string"
          },
          "telefone": {
            "type": "string"
          },
          "data_nascimento": {
            "type": "string"
          },
          "id_responsavel": {
            "type": "integer"
          }
        }
      },
      "DependenteArray": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/Dependente"
        }
      },
...
      "Endereco": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "logradouro": {
            "type": "string",
            "example": "Rua Alvarenga Dia"
          },
          "numero": {
            "type": "string",
            "example": "952 Casa C"
          },
          "cep": {
            "type": "string",
            "example": "00000-678"
          },
          "bairro": {
            "type": "string",
            "example": "Bairro Alv"
          },
          "cidade": {
            "type": "string",
            "example": "Belo Horizonte"
          },
          "estado": {
            "type": "string",
            "example": "MG"
          }
        }
      },
      "EnderecoPost": {
        "type": "object",
        "properties": {
          "logradouro": {
            "type": "string",
            "example": "Rua Alvarenga Dia"
          },
          "numero": {
            "type": "string",
            "example": "952 Casa C"
          },
          "cep": {
            "type": "string",
            "example": "00000-678"
          },
          "bairro": {
            "type": "string",
            "example": "Bairro Alv"
          },
          "cidade": {
            "type": "string",
            "example": "Belo Horizonte"
          },
          "estado": {
            "type": "string",
            "example": "MG"
          }
        }
      },
...
      "TokenInvalido": {
        "type": "object",
        "properties": {
          "cause": {
            "type": "string",
            "example": "Invalid token"
          },
          "message_error": {
            "type": "string",
            "example": "The Token has expired on 2023-10-16T18:43:02Z."
          },
          "path": {
            "type": "string",
            "example": "/api/cesta"
          }
        }
      },
      "DistribuicaoCesta": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 999
          },
          "data_hora": {
            "type": "string",
            "example": "2023-10-27T20:36:38.150664"
          },
          "beneficiario": {
            "$ref": "#/components/schemas/BeneficiarioDistribuicaoCesta"
          },
          "voluntario": {
            "$ref": "#/components/schemas/VoluntarioDistribuicaoCesta"
          },
          "cesta": {
            "$ref": "#/components/schemas/CestaDistribuicao"
          }
        }
      },
      "DistribuicaoCestaArray": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/DistribuicaoCesta"
        }
      },
      "DistribuicaoCestaPost": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 999
          },
          "data_hora": {
            "type": "string",
            "example": "2023-10-27T20:36:38.150664"
          },
          "beneficiario": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer",
                "example": 21
              }
            }
          },
          "voluntario": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer",
                "example": 244
              }
            }
          },
          "cesta": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer",
                "example": 233
              }
            }
          }
        }
      },
...
      "ResumoQuantidadesPorMesAno": {
        "type": "object",
        "properties": {
          "ano": {
            "type": "integer",
            "example": 2023
          },
          "mes": {
            "type": "integer",
            "example": 10
          },
          "quantidade": {
            "type": "integer",
            "example": 45
          }
        }
      },
      "ResumoQuantidadesPorMesAnoArray": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/ResumoQuantidadesPorMesAno"
        }
      }
    },
    "securitySchemes": {
      "auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}